src/communication/zmq_server.py
"""
ZeroMQ Communication Server Module

Implements ZeroMQ server for real-time communication with GUI.
Required by Section 8 of numerical_state_space_and_simulation_specification.md

Communication pattern: PUB-SUB for state broadcasting, REQ-REP for commands
"""

import zmq
import threading
import time


class ZMQServer:
    """
    ZeroMQ server for GUI communication.
    
    References:
    - numerical_state_space_and_simulation_specification.md: Section 8
    """
    
    def __init__(self, pub_port=5555, rep_port=5556):
        """
        Initialize ZeroMQ server.
        
        Args:
            pub_port: Port for publishing system state
            rep_port: Port for receiving commands
        """
        self.pub_port = pub_port
        self.rep_port = rep_port
        
        self.context = None
        self.pub_socket = None
        self.rep_socket = None
        
        self.running = False
        self.command_thread = None
    
    def start(self):
        """
        Start ZeroMQ server and communication threads.
        """
        pass
    
    def stop(self):
        """
        Stop ZeroMQ server and close sockets.
        """
        pass
    
    def publish_state(self, system_state):
        """
        Publish system state to GUI subscribers.
        
        Args:
            system_state: SystemState instance
        """
        pass
    
    def handle_commands(self):
        """
        Handle incoming commands from GUI (runs in separate thread).
        """
        pass
    
    def process_command(self, command_message):
        """
        Process received command message.
        
        Args:
            command_message: CommandMessage instance
        
        Returns:
            response: Response message
        """
        pass
