requirements.txt
# Python Dependencies for Industrial Pressure Control System Simulation
# 
# All dependencies required for implementation of the predefined system architecture.
# References: docs/numerical_state_space_and_simulation_specification.md

# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Plotting and visualization
matplotlib>=3.7.0

# Communication
pyzmq>=25.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Optional: For enhanced analysis
control>=0.9.0  # Control systems library (optional)
